Add: ViewFactory 파일 생성 후 이동
