+++ COMFIE/App/DIContainer/ViewFactory.swift
@@ -1,0 +1,32 @@
+//
+//  ViewFactory.swift
+//  COMFIE
+//
+//  Created by Anjin on 3/6/25.
+//
+
+import SwiftUI
+
+struct ViewFactory {
+    let router: Router
+    
+    // MARK: - Intent
+    private func makeOnboardingIntent() -> OnboardingStore { OnboardingStore(router: router) }
+    private func makeMemoIntent() -> MemoStore { MemoStore(router: router) }
+    
+    // MARK: - View
+    @ViewBuilder func makeView(_ route: Route) -> some View {
+        switch route {
+        case .loading:
+            LoadingView()
+        case .onboarding:
+            OnboardingView(intent: makeOnboardingIntent())
+        case .memo:
+            MemoView(intent: makeMemoIntent())
+        case .retrospection:
+            Text("retrospection")
+        case .comfieZoneSetting:
+            Text("comfieZoneSetting")
+        }
+    }
+}
