+++ COMFIE/App/Router/Route.swift
@@ -19,27 +19,3 @@ enum Route {
         ViewFactory(router: router).makeView(self)
     }
 }
-
-struct ViewFactory {
-    let router: Router
-    
-    // MARK: - Intent
-    private func makeOnboardingIntent() -> OnboardingStore { OnboardingStore(router: router) }
-    private func makeMemoIntent() -> MemoStore { MemoStore(router: router) }
-    
-    // MARK: - View
-    @ViewBuilder func makeView(_ route: Route) -> some View {
-        switch route {
-        case .loading:
-            LoadingView()
-        case .onboarding:
-            OnboardingView(intent: makeOnboardingIntent())
-        case .memo:
-            MemoView(intent: makeMemoIntent())
-        case .retrospection:
-            Text("retrospection")
-        case .comfieZoneSetting:
-            Text("comfieZoneSetting")
-        }
-    }
-}
